# openapi.yaml
openapi: 3.1.0
info:
  title: Terrakube Self-Service
  description: 'Self-service environments (**labs**) on Terrakube. A portal lists
    templates, the user fills in a form, and the

    service creates a Terrakube workspace, applies it, hands over its access details,
    and destroys it when its

    lifetime ends.


    **Authentication.** Every `/v1` call carries the portal''s API key (`Authorization:
    Bearer <key>`) and the end

    user: a verified OIDC ID token in `X-User-Token` (token mode) or `X-Actor-Email`
    set by the portal backend from its

    own session (header mode). Call the service from the portal backend only, never
    from a browser.


    **Roles.** `user`: own labs only (another user''s lab answers `404`). `auditor`:
    also reads every lab and the

    analytics, but cannot act on other people''s labs or read their access details
    (`403`). `admin`: everything.


    **Lifecycle.** `pending` → `provisioning` → `ready` or `failed` → `destroying`
    → `destroyed` or `destroy_failed`.

    Poll a lab every 15-30 s while it is `pending` or `provisioning`.


    **Lifetimes.** Labs expire at `expires_at` and are destroyed automatically; `extend`
    adds hours up to the

    template''s `max_ttl_hours` from creation.


    Integration guide: https://github.com/3sky/terrakube-selfservice/blob/main/docs/portal-integration.md'
  version: 0.6.1
paths:
  /v1/templates:
    get:
      tags:
      - templates
      summary: List lab templates
      operationId: list_templates
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateList'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /v1/templates/{template_id}:
    get:
      tags:
      - templates
      summary: Get a template and its form inputs
      operationId: get_template
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found, or not visible to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/templates/{template_id}/estimate:
    post:
      tags:
      - templates
      summary: Estimate a lab's cost before creating it
      description: Hourly and total list-price estimate for these form inputs; 404
        when the template has no cost model.
      operationId: estimate_template
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          title: Template Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estimate'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found, or not visible to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/labs:
    post:
      tags:
      - labs
      summary: Create a lab from a template
      description: Creates the Terrakube workspace and queues its apply. Poll the
        lab until `ready` or `failed`.
      operationId: create_lab
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Actor-Email
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Header mode: the signed-in user, set by the portal backend
            from its own session. Ignored in token mode.'
          title: X-Actor-Email
        description: 'Header mode: the signed-in user, set by the portal backend from
          its own session. Ignored in token mode.'
      - name: X-User-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Token mode: the signed-in user''s OIDC ID token, verified
            by the service.'
          title: X-User-Token
        description: 'Token mode: the signed-in user''s OIDC ID token, verified by
          the service.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: The user may not do this
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Conflict with the lab's current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '502':
          description: Terrakube or OpenBao request failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      tags:
      - labs
      summary: List labs
      operationId: list_labs
      security:
      - HTTPBearer: []
      parameters:
      - name: owner_email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Auditors and admins; users always see their own labs.
          title: Owner Email
        description: Auditors and admins; users always see their own labs.
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/LabStatus'
          - type: 'null'
          title: Status
      - name: template_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Id
      - name: include_destroyed
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Destroyed
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: X-Actor-Email
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Header mode: the signed-in user, set by the portal backend
            from its own session. Ignored in token mode.'
          title: X-Actor-Email
        description: 'Header mode: the signed-in user, set by the portal backend from
          its own session. Ignored in token mode.'
      - name: X-User-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Token mode: the signed-in user''s OIDC ID token, verified
            by the service.'
          title: X-User-Token
        description: 'Token mode: the signed-in user''s OIDC ID token, verified by
          the service.'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabList'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/labs/{lab_id}:
    get:
      tags:
      - labs
      summary: Get a lab
      operationId: get_lab
      security:
      - HTTPBearer: []
      parameters:
      - name: lab_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Lab Id
      - name: X-Actor-Email
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Header mode: the signed-in user, set by the portal backend
            from its own session. Ignored in token mode.'
          title: X-Actor-Email
        description: 'Header mode: the signed-in user, set by the portal backend from
          its own session. Ignored in token mode.'
      - name: X-User-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Token mode: the signed-in user''s OIDC ID token, verified
            by the service.'
          title: X-User-Token
        description: 'Token mode: the signed-in user''s OIDC ID token, verified by
          the service.'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found, or not visible to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/labs/{lab_id}/access:
    get:
      tags:
      - labs
      summary: Access details of a ready lab
      description: 'Kubeconfig, passwords, URLs published by the template. Owner or
        admin only (auditors get 403); every read is audited.


        The response is marked `Cache-Control: no-store`: show it to the user, never
        cache, log or store it.'
      operationId: lab_access
      security:
      - HTTPBearer: []
      parameters:
      - name: lab_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Lab Id
      - name: X-Actor-Email
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Header mode: the signed-in user, set by the portal backend
            from its own session. Ignored in token mode.'
          title: X-Actor-Email
        description: 'Header mode: the signed-in user, set by the portal backend from
          its own session. Ignored in token mode.'
      - name: X-User-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Token mode: the signed-in user''s OIDC ID token, verified
            by the service.'
          title: X-User-Token
        description: 'Token mode: the signed-in user''s OIDC ID token, verified by
          the service.'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabAccess'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: The user may not do this
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found, or not visible to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Conflict with the lab's current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '501':
          description: Not configured on this installation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '502':
          description: Terrakube or OpenBao request failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/labs/{lab_id}/extend:
    post:
      tags:
      - labs
      summary: Extend a lab's TTL
      operationId: extend_lab
      security:
      - HTTPBearer: []
      parameters:
      - name: lab_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Lab Id
      - name: X-Actor-Email
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Header mode: the signed-in user, set by the portal backend
            from its own session. Ignored in token mode.'
          title: X-Actor-Email
        description: 'Header mode: the signed-in user, set by the portal backend from
          its own session. Ignored in token mode.'
      - name: X-User-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Token mode: the signed-in user''s OIDC ID token, verified
            by the service.'
          title: X-User-Token
        description: 'Token mode: the signed-in user''s OIDC ID token, verified by
          the service.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '401':
          description: Missing or invalid API key or user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: The user may not do this
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found, or not visible to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Conflict with the lab's current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LabStatus:
      type: string
      enum:
      - pending
      - provisioning
      - ready
      - failed
      - destroying
      - destroy_failed
      - destroyed
      title: LabStatus
...
    Problem:
      properties:
        detail:
          anyOf:
          - type: string
          - items: {}
            type: array
          title: Detail
          description: Message, or a list of field errors for 422 validation failures.
      type: object
      required:
      - detail
      title: Problem
    Template:
      properties:
        id:
          type: string
          pattern: ^[a-z][a-z0-9-]*$
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ''
        default_ttl_hours:
          type: integer
          exclusiveMinimum: 0.0
          title: Default Ttl Hours
        max_ttl_hours:
          type: integer
          exclusiveMinimum: 0.0
          title: Max Ttl Hours
          description: Maximum lifetime from creation, including extensions.
        inputs:
          items:
            $ref: '#/components/schemas/TemplateInput'
          type: array
          title: Inputs
          default: []
      type: object
      required:
      - id
      - name
      - default_ttl_hours
      - max_ttl_hours
      title: Template
...
    TemplateList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Template'
          type: array
          title: Items
      type: object
      required:
      - items
      title: TemplateList
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: API key issued to the calling portal.
      scheme: bearer
...